+++ ips/services/__init__.py
@@ -1,0 +1,3 @@
+from ips.util.ui_configuration import UIConfiguration
+
+API_TARGET = UIConfiguration().get_api_uri()
