+++ ips/services/system_info.py
@@ -1,6 +1,6 @@
 from flask import request, render_template, Blueprint
 from flask_login import login_required
-from ips.persistence import app_methods
+from ips.services import app_methods
 
 bp = Blueprint('system_info', __name__, url_prefix='/system_info', static_folder='static')
 
@@ -8,7 +8,6 @@ bp = Blueprint('system_info', __name__, url_prefix='/system_info', static_folder
 @bp.route('/')
 @login_required
 def system_info():
-    print(request.method)
     records = app_methods.get_system_info()
     header = records[0]
     records = records[1:]
