+++ ips/templates/dashboard.html
@@ -28,7 +28,7 @@
                             <a href="{{ url_for('dashboard.dashboard_view') }}" class="nav__link">Dashboard</a>
                         </li>
                         <li class="nav__item " role="menuitem">
-                            <a href="{{  url_for('new_run.new_run_1')}}" class="nav__link">New Run</a>
+                            <a href="{{  url_for('new_run_steps.new_run_1')}}" class="nav__link">New Run</a>
                         </li>
                         <li class="nav__item " role="menuitem">
                             <a href="{{ url_for('system_info.system_info')}}" class="nav__link">System Info</a>
